+++ src/main/java/wooteco/subway/service/PathService.java
@@ -3,16 +3,11 @@ package wooteco.subway.service;
 import static java.util.stream.Collectors.toList;
 
 import java.util.List;
-import java.util.Objects;
-import org.jgrapht.GraphPath;
-import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
-import org.jgrapht.graph.DefaultWeightedEdge;
-import org.jgrapht.graph.WeightedMultigraph;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.dao.SectionDao;
 import wooteco.subway.domain.Fare;
-import wooteco.subway.domain.Section;
+import wooteco.subway.domain.Path;
 import wooteco.subway.domain.Sections;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.PathResponse;
@@ -34,41 +29,20 @@ public class PathService {
         Station source = stationService.findById(sourceId).toStation();
         Station target = stationService.findById(targetId).toStation();
 
-
-        WeightedMultigraph<Station, DefaultWeightedEdge> graph = new WeightedMultigraph(DefaultWeightedEdge.class);
-
         Sections sections = new Sections(sectionDao.findAll());
-        for (Station station : sections.getStations()) {
-            graph.addVertex(station);
-        }
+        Path shortestPath = new Path(sections);
 
-        for (Section section : sections.getValue()) {
-            graph.setEdgeWeight(graph.addEdge(section.getUpStation(), section.getDownStation()), section.getDistance());
-        }
+        List<StationResponse> stationResponses = convertToStationResponse(source, target, shortestPath);
 
-        DijkstraShortestPath dijkstraShortestPath = new DijkstraShortestPath(graph);
-        GraphPath<Station, DefaultWeightedEdge> shortestPath = getPath(source, target, dijkstraShortestPath);
-
-        if (Objects.isNull(shortestPath)) {
-            throw new IllegalArgumentException("경로가 존재하지 않습니다.");
-        }
-
-        List<Station> vertexes = shortestPath.getVertexList();
-
-        List<StationResponse> stationResponses = vertexes.stream()
-                .map(StationResponse::new)
-                .collect(toList());
-
-        int shortestDistance = (int) shortestPath.getWeight();
+        int shortestDistance = shortestPath.getDistance(source, target);
         Fare fare = new Fare(shortestDistance);
         return new PathResponse(stationResponses, shortestDistance, fare.calculate());
     }
 
-    private GraphPath getPath(Station source, Station target, DijkstraShortestPath dijkstraShortestPath) {
-        try {
-            return dijkstraShortestPath.getPath(source, target);
-        } catch (IllegalArgumentException e) {
-            throw new IllegalArgumentException("노선에 등록되지 않는 지하철역입니다.");
-        }
+    private List<StationResponse> convertToStationResponse(Station source, Station target, Path shortestPath) {
+        return shortestPath.getStations(source, target)
+                .stream()
+                .map(StationResponse::new)
+                .collect(toList());
     }
 }
